+++ modulo4/to-do-list/src/index.ts
@@ -1,21 +1,7 @@
 import express, { Express, Request, Response } from "express";
-import knex from "knex";
 import cors from "cors";
-import dotenv from "dotenv";
+import { connection } from "./connections";
 import { AddressInfo } from "net";
-
-dotenv.config();
-
-export const connection = knex({
-	client: "mysql",
-	connection: {
-    host: process.env.DB_HOST,
-    port: 3306,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASS,
-    database: process.env.DB_NAME
-  }
-});
 
 const app: Express = express();
 app.use(express.json());
@@ -34,14 +20,6 @@ const server = app.listen(process.env.PORT || 3003, () => {
 app.post("/user", async (req: Request, res: Response)=>{
     try{
         await connection("Actor")
-        // .update({
-        // name: req.body.name,
-        // salary: req.body.salary,
-        // birth_date: req.body.birthDate,
-        // gender: req.body.gender
-        // })
-        // .where({ id: req.params.id })
-        
         res.status(200).send({chaveDoRetorno: valorDaBusca});
     }catch(error){
         res.status(400).send({chaveDoErro: valorDoErro});
